tests/wpt/tests/html/dom/render-blocking/element-render-blocking-007.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Media attribute that doesn't match makes the link not apply</title>

<link rel=expect href="#last" blocking="render" media="(max-width: 10px)">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "link with non-matching media has no effect");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
